Create new test suite for windows compatibility tests
